Restructure package structure
